refactor: move voting power and hive extras to hive module for better organization
