+++ src/v4vapp_backend_v2/models/hive_models.py
@@ -19,7 +19,7 @@ from v4vapp_backend_v2.helpers.crypto_conversion import CryptoConv, CryptoConver
 # )
 from v4vapp_backend_v2.helpers.crypto_prices import AllQuotes, QuoteResponse
 from v4vapp_backend_v2.helpers.general_purpose_funcs import seconds_only
-from v4vapp_backend_v2.helpers.hive_extras import (
+from v4vapp_backend_v2.hive.hive_extras import (
     HiveTransactionTypes,
     decode_memo,
     get_blockchain_instance,
